fixes needed to correctly inspect remote images

+++ acceptance/invoke/pack.go
@@ -63,6 +63,9 @@ func NewPackInvoker(
 }
 
 func (i *PackInvoker) Cleanup() {
+	if i == nil {
+		return
+	}
 	i.testObject.Helper()
 
 	err := os.RemoveAll(i.home)
